build: make it compatible with python3 tools
sphinx-build and rst2man binaries on Fedora 28 have a -3 suffix when
they are installed for python3 in order to be able to be installed in
parallel to the python2's package.

For python3-sphinx, there's a bash module that plugs into /etc/profile.d
to add /usr/libexec/python3-sphinx to PATH. That however doesn't work
if you don't reload the profile after you installed the package.

So this makes it simpler by stating that we are compatible with any
of the tools (and we prefer the one with -3 suffix if available).
While at it, also allow people to override RST2MAN in the command line
in case we have even more creative scenarios.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>