dim: Remove upgrade logic
Of course I forgot the imporant part.

Note: Never push this, or anything after it, to the old place in
drm-tip/maintainer-tools.

Given our track record of forgetting a critical part in migrations I
think it'd be good if we hold off for a week or so until we apply this
one though. It's just a few wasted cpu cycles when running dim
update-branches.

Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>