dim: Add range-diff convenience wrapper
range-diff is awesome, but the interface is a bit silly. Add a bunch
of shortcuts, inspired by what git diff does.

v2: Add it to the developer commmands list. With this dim range-diff
is useable on any git repo, not just a dim managed one.

v3: Comments from Lucas:
- Forward additional arguments. There's no --no-patch in upstream git,
  but I tested this with --no-color and --no-dual-color. Slightly a
  hack since detecting whether it's a branch or not became even more
  fun, but works in all cases.
- Remove impossible "$branch" == "" check - I added a default.

v4: Update docs.

Cc: Lucas De Marchi <lucas.de.marchi@gmail.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>