dim: make apply-patch fail again
Somewhere since I originally typed this and today git apply-mbox lost
the error status for "this patch didn't apply cleanly". So instead of
exiting, we happily continue on and add Link: tags and other nonsense
to the current patch, instead of the one that we wanted to apply.

Fix this.

v2: It wasn't git am, it was us, wrapping the apply_patch function in
an if, which defeats set -e. Add explicit error handling. Thanks to
Jani for figuring this one out.

Cc: Jani Nikula <jani.nikula@linux.intel.com>
Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>