dim: Sort sparse output after remapping
If patch happens to move lines around it's easy to end in a situation
where we end up with errors just moving around:

-error1:123
 error2:124
+error1:125

With remapping, sorting and then diffing without context we end up with:
-error1:123
+error1:123

Which is quite confusing and should not generate warning in the first
place.

This can be fixed by sorting the lines *after* the remapping.

Cc: Daniel Vetter <daniel@ffwll.ch>
Reported-by: Oleg Vasilev <oleg.vasilev@intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>