dim: Don't freeze sparse check when there are Kconfig changes
If the commit range that is being checked with 'dim sparse' introduces
new Kconfig options 'make' prompts us for our opinion. Since we are
capturing the stdout and -err the user sees nothing and we are seemingly
stuck.

To avoid that and be as non-interactive as possible let's just squeeze
'make olddefconfig' in-between 'git checkout' and the actual 'make'.

Reported-by: Jani Nikula <jani.nikula@linux.intel.com>
Signed-off-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>