dim: Create git_push helper
Just a bit of ocd ... One functional change is the switch from $DRY to
the $DRY_RUN flag in dim_pull_request. That one didn't have a
justification, and the one for the other place in dim_create_branch
didn't seem to apply really.

v2: don't substitute too much (Arek)

Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>