dim: Make sure our local branch is pushed before creating a tag
The current dim code will create the tag using a local branch as an
argument, but while we make sure to fetch the branch first in the callers,
we don't actually make sure that any local change have been pushed.

Make sure it's the case before creating the tag by checking that all local
commits have been pushed to the meaningful remote.

Signed-off-by: Maxime Ripard <mripard@kernel.org>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>