dim: Handle multiple From: lines in pull request
dim blew up on the pull from Thomas:

https://patchwork.freedesktop.org/patch/348981/

Thomas helpfully added an in-body From: line, similar to how git
format-patch does. But our dim parsing choked on the double From: line
that resulted in.

Assume that if there's a 2nd From: line it's there for a good reason
and pick that one (like git apply-mbox).

v2: Ditch the debug stuff ... silly me.

Cc: "Thomas Hellström (VMware)" <thomas_os@shipmail.org>
Cc: Dave Airlie <airlied@gmail.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reported-by: Dave Airlie <airlied@gmail.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>