dim: add a simple facility to require minimum dim version
Add simple single-digit dim version scheme, and add 'dim version'
subcommand to print it. Start off with version 0.

If $DIM_MIN_VERSION is set, bail out if the requirement is not met. The
idea is to set this in nightly.conf to be able to soft-enforce dim
updates before rolling out changes. (It's obviously possible to bypass
the requirement.)

We do have the dim uptodate nag, but it never bails out.

Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>