dim: change the drm-intel-next and drm-intel-next-queued process
Remove the special casing of drm-intel-next/drm-intel-next-queued
tagging and pull requests. Instead of having a two-step process for
dinq, follow the same pull request process for all branches. As the
first step, use drm-intel-next-queued for pull requests, and ignore
drm-intel-next for now. The change is only relevant for maintainers.

In the future, we'll want to switch from drm-intel-next-queued to
drm-intel-next completely. This will need a flag day switch for all
committers and maintainers.

We also decouple driver date update from dinq pull requests. We may want
to apply the driver date update to drm-intel-gt-next branch instead.

With the change, we also lose the ability to add multiple drm-intel-next
tags, and create a single pull request for them. In practice, we haven't
really used this possibility, as it lacks the merging of the annotated
tag changelogs together, missing information. We may want to resurrect
the feature in an improved form in the future, generalized for all
branches.

Cc: Daniel Vetter <daniel@ffwll.ch>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>