dim: Error checking for git commit in rebuild-tip
There's nothing we can do, but also I was baffled about what's going on.
The reason was that a topic branch was deleted on the server, but not
removed from nightly.conf. Which means git merge fails, but not in a way
we expect, so we keep falling over until this git commit (which should
always succeed) also falls over and the scrip just stops.

Make the failure a notch more verbose for next time around.

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>