doc: Mention dim cherry-pick command in getting started page
The dim tool supports a cherry-pick that is a glorified version of the git
command but adds additional metadata and checks if there are any follow-up
commits referencing the one being cherry-picked.

Let's document in the getting started page that the dim command should be
used instead of just git cherry-pick. That way people will be aware of it.

This is useful to avoid cherry-picking commits but miss existing fixes for
them, which would introduce unnecessary regressions in the target branch.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>
Acked-by: Thomas Zimmermann <tzimmermann@suse.de>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>