dim: bump dim version to 2
Commit ea3d1418a434 ("dim: Use git apply to apply patch") changed dim
behaviour in that it now accepts fixup patches that can create new
files, and those need to be applied using 'git apply' instead of plain
'patch'. Bump dim version to 2 to be able to require this feature.

We'll also need to bump the dim version requirement (DIM_MIN_VERSION) to
2 in nightly.conf to account for the fixups that need this change.

Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Cc: Sam Ravnborg <sam@ravnborg.org>
Cc: Thomas Zimmermann <tzimmermann@suse.de>
Signed-off-by: Jani Nikula <jani.nikula@intel.com>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>