dim: fix From: matching in apply patch
I occasionally bounce patches to intel-gfx@ to get CI for patches that
were only sent to dri-devel@ or were caught in the moderation
queue. This results in a Resent-From: header being added.

Applying such a patch leads to the From: match in apply_patch() to match
Resent-From: rather than the actual From:. If whoever bounced the patch
is also applying the patch, their Signed-off-by won't be automatically
added.

Match From: from the beginning of the line to fix this.

Signed-off-by: Jani Nikula <jani.nikula@intel.com>