dim_setup: Allow to use other setup as reference
On a slow network, most of the time for a setup is taken by the initial
clone of the kernel repository. It's common for kernel developers to
have a kernel clone lying around, so allow to borrow the git objects
from there.

Testing on a relatively fast connection, the initial checkout goes down
from ~7m45s to 2m21s.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>