dim: Pull requests to use https://
Not all repos have a git:// protocol enabled. Once upon a time the
justification to use it was due to bad performance of the https backend.
https://git-scm.com/book/en/v2/Git-on-the-Server-The-Protocols still
marks it as "the best on performance", but with the Smart HTTP backend,
the cons start to have a bigger role. Also worth noting that it's not
available on Github and Gitlab anymore due to security issues.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>