dim: Check for maintainer's ack when merging on behalf of another tree
We've recently had multiple occurences of people taking patches through
the drm-misc tree for parts of the kernel that aren't under drm-misc
maintenance, and without the maintainers Ack.

While merging patches on behalf of someone else is totally something we
can do, we need to tell the maintainer and make sure he's fine with
that.

This patch adds some logic to detect if the the range of commits belong
indeed to the branch we are currently working on, and if not check for
the maintainer's Acked-by or Signed-off-by.

Signed-off-by: Maxime Ripard <mripard@kernel.org>