dim: Allow pull-request from previously created tag
It may be desired to tag separately from the pull request creation.
Either because we want more tests to happen on the tag or because
we want to tag something that is not the tip of the branch. The latter
is useful if there were last minute pushes to -next and it still didn't
receive the proper amount of testing in linux-next to be part of a pull
request.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>