dim: Fix default drm-tip remote name for new setup
We used to call the remote drm-tip. However since the move to gitlab,
and since that in a new setup the integration manifest is still not
available, the drm-tip remote is now called "tip" by default. That may
be unexpected, particularly when setting up with --yes since we are not
prompted for a different name.

Change the logic in url_to_remote to deal better with the bootstrapping
phase: it should only ever happen with drm-tip. No other case should
need the fallback to the remote name, as those are derived from
the manifest.

Also move url assignment earlier so the url suggested is always the
first, not the last.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>