Revert "dim: tag once"
The -e usage breaks EDITOR="emacs -nw":

hint: Waiting for your editor to close the file... emacs: standard input is not a tty
error: There was a problem with the editor 'emacs -nw'.
Please supply the message using either -m or -F option.

And simply removing the -e tag didn't worked out. It then
simply goes there and create the tag with the template without
stopping on emacs.

This reverts commit b84094292a9d7f27d9f2161ce21757f4eb1a2948.