dim: Drop fetch in dim_cherry_pick
The fetch is currently called with an unbound `remote` variable,
which results in fetching the upstream remote of the current branch.
In most cases this is the correct behavior as the branch we are
cherry-picking changes in should have the upstream set. However if
the upstream is not set, it will a fetch origin, which has nothing
to do with the operation.

Just drop the fetch as it's not needed for 2 reasons:

1) If the user is giving commit X, he probably have that commit locally
   already and no fetch is needed
2) When listing the references, we will fetch drm-tip, which should have
   all commits we are possibly cherry-picking

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>