doc: clean up access-request documentation
Unify access request text for all the repos. This changes the text
slightly for all repos:

- drm-misc gains "Commit rights may be automatically revoked after a
  year of inactivity (no commits or reviews)". For drm-intel/drm-xe that
  changes from "will be" to "may be" revoked.

- drm-intel/drm-xe lose the text on rate limiting the number of new
  committers. This was barely an issue when we started adding committers
  in the first place, but certainly not anymore.

- Remove drm-intel/drm-xe repo details. They belong in repo
  documentation and MAINTAINERS.

Add links to creating access request issues to drm-intel/drm-xe
too. They URLs contain the info about the issue template to use, which
we don't have in place yet, but there's no harm in using that already.

Signed-off-by: Jani Nikula <jani.nikula@intel.com>