doc: add redirects for moved pages
When moving pages around I, for whatever reason, didn't consider that
this creates a bunch of stale links. The drm-misc link in MAINTAINERS
being perhaps the worst one to go 404.

We could, and should, fix a lot of the now stale links, but in the mean
time, use the sphinx-reredirects extension [1] to create redirects for
the relocated pages.

[1] https://documatt.com/sphinx-reredirects/

Signed-off-by: Jani Nikula <jani.nikula@intel.com>