dim: Allow passing rebase options to retip with implicit branch
The subcommand "retip" allows passing extra git rebase options, but that
requires the branch name to be explicitly passed as first argument.
Since we do not expect branch names to start with a dash, let's allow
the user to pass extra rebase options without the need to explicitly
pass the branch name if the current one is to be used.

Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>