dim: Fix dim setup without a reference
Tejas reported that dim fails on a simple `dim setup` with the following
message:

	The DIM_REPO maintainer kernel repository 'src' doesn't exist. Clone upstream? (y/N) y
	Cloning into 'src'...
	info: Could not add alternate for '': path '' does not exist

git should continue if the path doesn't exist, but apparently this
doesn't happen with all git versions if the argument is an empty string.
The original commit tried to workaround that by setting it to a
non-existent path, but there was a typo in the variable name. Fix the
typo.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>