dim: Restore fix_of_fix
The original fix of fix on commit 281550897a58 ("dim: Consider
fix-of-fix for -fixes cherry picking.") was:

fix_of_fix=$(git log --grep="cherry picked from commit $fixes" \
		  --after=6months --format=format:%h \
		  $intel_remote/drm-intel-next-fixes -1)
if [[ -z "$fix_of_fix" ]]; then
 fix_of_fix=$(git log --grep="cherry picked from commit $fixes" \
 --after=6months --format=format:%h $intel_remote/drm-intel-fixes -1)

But commit e7f730565356 ("dim: Generalize cherry-pick fixes")
generalized it to for b in drm-intel-next drm-intel-gt-next

So, we need to get back to look for the commits in the fix branch
not in the next branch. However, let's only look to the current
fix branch instead of both fixes branches, otherwise we will can
reach a point where we can try to cherry-pick a fix_of_fix to a branch
which doesn't have the first-fix patch.

Cc: Lucas De Marchi <lucas.demarchi@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>