dim: Better bash completion integration
I started noticing some programs that aren't integrated as a distro
package to have a command line switch to dump the bash-completion script
so it can be sourced, without the user worrying where it was installed.

Example: https://fx.wtf/install

Add the equivalent for dim so the autocompletion can be more easily
added to dev's workflow.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>