dim: mitigate get_maintainer.pl failing for added files
When the commit has non-empty new files, get_maintainer.pl doesn't seem
to order the URLs the way we expect:

New empty file seems ok:
	$ touch drivers/gpu/drm/xe/test_maint.c
	$ git add drivers/gpu/drm/xe/test_maint.c
	$ git commit -m tmp
	$ git show --no-use-mailmap --pretty=email HEAD | ./scripts/get_maintainer.pl --no-email --no-l --scm
	git https://gitlab.freedesktop.org/drm/xe/kernel.git
	git https://gitlab.freedesktop.org/drm/kernel.git
	git git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git

New non-empty file has Linus' tree as the first one:
	$ echo " " > drivers/gpu/drm/xe/test_maint2.c
	$ git add drivers/gpu/drm/xe/test_maint2.c
	$ git commit -m tmp2
	$ git show --no-use-mailmap --pretty=email HEAD | ./scripts/get_maintainer.pl --no-email --no-l --scm
	git git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
	git https://gitlab.freedesktop.org/drm/xe/kernel.git
	git https://gitlab.freedesktop.org/drm/kernel.git

It shows LKML as the first URL and then the check fail due to missing
Ack, even if the added file is within the dir that is maintained by the
tree being pushed to.

Instead of reverting commit 43254c2aa575 ("dim: Check for maintainer's
ack when merging on behalf of another tree"), just mitigate it for added
files. There's probably something to adapt in get_maintainer.pl to
completely fix it.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>