drm-intel: Patch flow updated
This picture was outdated. Update it with a new version that
includes drm-intel-gt-next and drm-xe-* branches.

Also remove the rc timeline to make it easier to generated
and to read.

The plantuml was created with help from AI, but manually cared,
while the graphiz was a blind AI conversion.

A Graphiz conversion was attempted, but the result of the
plantuml was more appealing and easier to read in the browser.

v2: - squash both plantuml and graphiz instead of intermidiate
      conversion (Lucas)
    - Typo and text adjustments (Lucas)
    - gitlab deps (Lucas)
v3: - Rebase
    - Add python3-pip dep to gitlab-ci

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>